Tile implementation(cont.)

+++ Assets/Script/GridMap.cs
@@ -4,20 +4,18 @@ using UnityEngine;
 
 public class GridMap : MonoBehaviour
 {
-    private int length;
-    private int width;
-    private int[,] gridMapArray;
+    private float length;
+    private float width;
+    private float[,] gridMapArray;
 
     public GridMap(int width, int length)
     {
         this.width = width;
         this.length = length;
 
-        gridMapArray = new int[width, length];
+        gridMapArray = new float[width, length];
 
         Debug.Log(width + " " + length);
-
-
     }
 
 }
